elecron: add annotations from all windows
