+++ packages/electron/src/main/attributes/WindowAttributeProvider.ts
@@ -1,5 +1,6 @@
 import { BacktraceAttributeProvider } from '@backtrace/sdk-core';
-import { BrowserWindow, WebFrameMain } from 'electron';
+import { BrowserWindow } from 'electron';
+import { flatten, getBasicBrowserWindowAttributes } from './helpers/attributes';
 
 export interface WindowAttributeProviderOptions {
     /**
@@ -8,16 +9,6 @@ export interface WindowAttributeProviderOptions {
      * The keys will take form of `electron.window.${key}.*`.
      */
     readonly key?: string;
-
-    /**
-     * If `true`, attributes will be added. Default: `true` if `key` is provided, otherwise `false`.
-     */
-    readonly attributes?: boolean;
-
-    /**
-     * If `true`, annotations will be added. Default: `true`.
-     */
-    readonly annotations?: boolean;
 
     /**
      * If `true`, destroyed windows will not add any information. Default: `true`.
@@ -31,192 +22,12 @@ export class WindowAttributeProvider implements BacktraceAttributeProvider {
     public readonly type = 'dynamic';
 
     public get(): Record<string, unknown> {
-        const { key, attributes, annotations, includeDestroyed } = this._options ?? {};
+        const { key, includeDestroyed } = this._options ?? {};
 
         if (!includeDestroyed && this._window.isDestroyed()) {
             return {};
         }
 
-        const result: Record<string, unknown> = {};
-
-        if (annotations ?? true) {
-            result[`electron.window.${key ?? this._window.id}`] = this.getWindowObject(this._window);
-        }
-
-        if (attributes ?? !!key) {
-            Object.assign(
-                result,
-                this.flatten(this.getBasicWindowObject(this._window), `electron.window.${key ?? this._window.id}`),
-            );
-        }
-
-        return result;
-    }
-
-    private getBasicWindowObject(window: BrowserWindow) {
-        const parent = window.getParentWindow();
-
-        return {
-            accessibleTitle: window.accessibleTitle,
-            autoHideMenuBar: window.autoHideMenuBar,
-            closable: window.closable,
-            documentEdited: window.documentEdited,
-            excludedFromShownWindowsMenu: window.excludedFromShownWindowsMenu,
-            focusable: window.focusable,
-            fullScreen: window.fullScreen,
-            fullScreenable: window.fullScreenable,
-            id: window.id,
-            kiosk: window.kiosk,
-            maximizable: window.maximizable,
-            menuBarVisible: window.menuBarVisible,
-            minimizable: window.minimizable,
-            movable: window.movable,
-            representedFilename: window.representedFilename,
-            resizable: window.resizable,
-            shadow: window.shadow,
-            simpleFullScreen: window.simpleFullScreen,
-            title: window.title,
-            visibleOnAllWorkspaces: window.visibleOnAllWorkspaces,
-
-            isAlwaysOnTop: window.isAlwaysOnTop(),
-            isEnabled: window.isEnabled(),
-            isFocused: window.isFocused(),
-            isHiddenInMissionControl: window.isHiddenInMissionControl(),
-            isMaximized: window.isMaximized(),
-            isMenuBarAutoHide: window.isMenuBarAutoHide(),
-            isMinimized: window.isMinimized(),
-            isModal: window.isModal(),
-            isNormal: window.isNormal(),
-            isSimpleFullScreen: window.isSimpleFullScreen(),
-            isTabletMode: window.isTabletMode(),
-            isVisible: window.isVisible(),
-            isVisibleOnAllWorkspaces: window.isVisibleOnAllWorkspaces(),
-
-            backgroundColor: window.getBackgroundColor(),
-            bounds: window.getBounds(),
-            contentBounds: window.getContentBounds(),
-            contentSize: window.getContentSize(),
-            maximumSize: window.getMaximumSize(),
-            mediaSourceId: window.getMediaSourceId(),
-            minimumSize: window.getMinimumSize(),
-            normalBounds: window.getNormalBounds(),
-            opacity: window.getOpacity(),
-            position: window.getPosition(),
-            size: window.getSize(),
-            windowButtonPosition: window.getWindowButtonPosition(),
-
-            childWindows: window.getChildWindows().map((w) => this.getRef('windowId', w.id)),
-            parentWindow: parent ? this.getRef('windowId', parent.id) : null,
-
-            webContents: {
-                audioMuted: window.webContents.audioMuted,
-                backgroundThrottling: window.webContents.backgroundThrottling,
-                frameRate: window.webContents.frameRate,
-                id: window.webContents.id,
-                userAgent: window.webContents.userAgent,
-                zoomFactor: window.webContents.zoomFactor,
-                zoomLevel: window.webContents.zoomLevel,
-
-                isAudioMuted: window.webContents.isAudioMuted(),
-                isBeingCaptured: window.webContents.isBeingCaptured(),
-                isCrashed: window.webContents.isCrashed(),
-                isCurrentlyAudible: window.webContents.isCurrentlyAudible(),
-                isDestroyed: window.webContents.isDestroyed(),
-                isDevToolsFocused: window.webContents.isDevToolsFocused(),
-                isDevToolsOpened: window.webContents.isDevToolsOpened(),
-                isFocused: window.webContents.isFocused(),
-                isLoading: window.webContents.isLoading(),
-                isLoadingMainFrame: window.webContents.isLoadingMainFrame(),
-                isOffscreen: window.webContents.isOffscreen(),
-                isPainting: window.webContents.isPainting(),
-                isWaitingForResponse: window.webContents.isWaitingForResponse(),
-
-                mainFrame: this.getBasicFrameObject(window.webContents.mainFrame),
-                opener: window.webContents.opener
-                    ? this.getRef('frameTreeNodeId', window.webContents.opener.frameTreeNodeId)
-                    : null,
-            },
-        };
-    }
-
-    private getWindowObject(window: BrowserWindow, visited = new Set<number>()): object {
-        if (visited.has(window.id)) {
-            return this.getRef('windowId', window.id);
-        }
-
-        visited.add(window.id);
-
-        const parent = window.getParentWindow();
-        const basic = this.getBasicWindowObject(window);
-
-        return {
-            ...basic,
-
-            childWindows: window.getChildWindows().map((w) => this.getWindowObject(w, visited)),
-            parentWindow: parent ? this.getWindowObject(parent, visited) : null,
-
-            webContents: {
-                ...basic.webContents,
-
-                mainFrame: this.getFrameObject(window.webContents.mainFrame),
-            },
-        };
-    }
-
-    private getFrameObject(frame: WebFrameMain, visited = new Set<number>()): object {
-        if (visited.has(frame.frameTreeNodeId)) {
-            return this.getRef('frameTreeNodeId', frame.frameTreeNodeId);
-        }
-
-        visited.add(frame.frameTreeNodeId);
-        const basic = this.getBasicFrameObject(frame);
-
-        return {
-            ...basic,
-            framesInSubtree: frame.framesInSubtree.map((f) => this.getFrameObject(f, visited)),
-            top: frame.top ? this.getFrameObject(frame.top, visited) : null,
-        };
-    }
-
-    private getBasicFrameObject(frame: WebFrameMain): object {
-        return {
-            frameTreeNodeId: frame.frameTreeNodeId,
-            framesInSubtree: frame.framesInSubtree.map((f) => this.getRef('frameTreeNodeId', f.frameTreeNodeId)),
-            top: frame.top ? this.getRef('frameTreeNodeId', frame.top.frameTreeNodeId) : null,
-
-            name: frame.name,
-            origin: frame.origin,
-            osProcessId: frame.osProcessId,
-            routingId: frame.routingId,
-            url: frame.url,
-            visibilityState: frame.visibilityState,
-        };
-    }
-
-    private getRef(name: string, value: unknown) {
-        const ref = { [`$${name}`]: value };
-
-        return {
-            _ref: true,
-            ...ref,
-            toJSON() {
-                return ref;
-            },
-        };
-    }
-
-    private flatten(obj: object, parentKey?: string) {
-        const result: Record<string, unknown> = {};
-
-        for (const key in obj) {
-            const value = obj[key as keyof typeof obj];
-            if (typeof value === 'object') {
-                Object.assign(result, this.flatten(value, parentKey ? `${parentKey}.${key}` : key));
-            } else {
-                result[parentKey ? `${parentKey}.${key}` : key] = value;
-            }
-        }
-
-        return result;
+        return flatten(getBasicBrowserWindowAttributes(this._window), `electron.window.${key ?? this._window.id}`);
     }
 }
