+++ packages/electron/src/main/builder/BacktraceClientBuilder.ts
@@ -1,10 +1,11 @@
-import { BacktraceClientBuilder as NodeBacktraceClientBuilder, BacktraceNodeClientSetup } from '@backtrace/node';
+import { BacktraceNodeClientSetup, BacktraceClientBuilder as NodeBacktraceClientBuilder } from '@backtrace/node';
+import { BacktraceClient } from '../BacktraceClient';
+import { AllWindowsAttributeProvider } from '../attributes/AllWindowsAttributeProvider';
 import { AppAttributeProvider } from '../attributes/AppAttributeProvider';
 import { ApplicationInformationAttributeProvider } from '../attributes/ApplicationInformationAttributeProvider';
 import { GpuAttributeProvider } from '../attributes/GpuAttributeProvider';
 import { GpuFeatureAttributeProvider } from '../attributes/GpuFeatureAttributeProvider';
 import { ReadyAppAttributeProvider } from '../attributes/ReadyAppAttributeProvider';
-import { BacktraceClient } from '../BacktraceClient';
 
 export class BacktraceClientBuilder extends NodeBacktraceClientBuilder {
     constructor(clientSetup: BacktraceNodeClientSetup) {
@@ -15,6 +16,7 @@ export class BacktraceClientBuilder extends NodeBacktraceClientBuilder {
         this.addAttributeProvider(new GpuFeatureAttributeProvider());
         this.addAttributeProvider(new ReadyAppAttributeProvider());
         this.addAttributeProvider(new AppAttributeProvider());
+        this.addAttributeProvider(new AllWindowsAttributeProvider());
     }
 
     public build(): BacktraceClient {
